Add documentation to components and API functions
In this commit, we have added documentation in form of comments to the DeliverablesList, DeliverableDetail components and API method functions to improve code readability. Each of the comments added briefly describe what its respective function or component does to provide a clear and quick understanding of the code for future reference and maintenance activities.